+++ src/Products/PlonePAS/tools/memberdata.py
@@ -4,6 +4,7 @@ from AccessControl.requestmethod import postonly
 from Acquisition import aq_base
 from App.class_init import InitializeClass
 from Products.BTreeFolder2.BTreeFolder2 import BTreeFolder2
+from Products.CMFCore.MemberDataTool import _marker
 from Products.CMFCore.MemberDataTool import MemberData as BaseMemberData
 from Products.CMFCore.MemberDataTool import MemberDataTool as BaseTool
 from Products.CMFCore.permissions import ManagePortal
@@ -22,8 +23,6 @@ from Products.PluggableAuthService.interfaces.plugins import IPropertiesPlugin
 from Products.PluggableAuthService.interfaces.plugins import \
     IRoleAssignerPlugin
 from zope.interface import implementer
-
-_marker = object()
 
 
 class MemberDataTool(BaseTool):
@@ -302,7 +301,13 @@ class MemberData(BaseMemberData):
             # we won't always have PlonePAS users, due to acquisition,
             # nor are guaranteed property sheets
             if not sheets:
-                return BaseMemberData.getProperty(self, id, default)
+                try:
+                    return BaseMemberData.getProperty(self, id, default)
+                except ValueError:
+                    # Zope users don't have PropertySheets,
+                    # return an empty string for them if the property
+                    # doesn't exists.
+                    return ''
 
         # If we made this far, we found a PAS and some property sheets.
         for sheet in sheets: